Changed port number in .thrift file to be 16 bit.

+++ diqube-remote/diqube-remote-base/src/main/thrift/base.thrift
@@ -33,7 +33,7 @@ struct RNodeHttpAddress {
 
 struct RNodeDefaultAddress {
   1: string host,
-  2: i32 port
+  2: i16 port
 }
 
 enum RNodeAccessType {
